@@ -19,13 +19,12 @@ classifiers = [
   # https://en.wikipedia.org/wiki/History_of_Python#Support
   # https://endoflife.date/python
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 
 [tool.setuptools.dynamic]
 version = {attr = "plextraktsync.__version__"}
